+++ backend/src/main/java/com/example/demo/global/security/UserPrincipal.java
@@ -1,0 +1,69 @@
+package com.example.demo.global.security;
+
+import com.example.demo.domain.user.entity.User;
+import com.example.demo.domain.user.entity.UserStatus;
+import com.example.demo.domain.user.role.Role;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.UUID;
+
+@Getter
+public class UserPrincipal implements UserDetails {
+
+    private final UUID id;
+    private final String email;
+    private final String password;
+    private final Collection<? extends GrantedAuthority> authorities;
+    private final UserStatus status;
+    private final Role role;
+
+    // 생성자: User 엔티티를 받아서 UserDetails로 필요한 정보만 가공합니다.
+    public UserPrincipal(User user) {
+        this.id = user.getId();
+        this.email = user.getEmail();
+        this.password = user.getPassword();
+        this.authorities = Collections.singletonList(new SimpleGrantedAuthority(user.getRole().name()));
+        this.status = user.getStatus();
+        this.role = user.getRole();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return this.status != UserStatus.BANNED;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return this.status == UserStatus.ACTIVE;
+    }
+}
